addded url validation, repaired two last things: link data does not l…

+++ src/main/java/projekt/wsb/linkbinder/service/LinkServiceImpl.java
@@ -1,5 +1,6 @@
 package projekt.wsb.linkbinder.service;
 
+import org.apache.commons.validator.routines.UrlValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
@@ -15,7 +16,7 @@ import projekt.wsb.linkbinder.users.UserEntity;
 import java.util.Comparator;
 import java.util.List;
 import java.util.ListIterator;
-import java.util.Locale;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -29,10 +30,27 @@ class LinkServiceImpl implements LinkService {
     @Override
     public String deleteLink(String currentLink, String loggedUsername, Model model) {
 
-        String tablename = linkRepository.findById(currentLink).get().getTableEntity().getTablename();
-        linkRepository.deleteById(currentLink);
-        addModelAttributes(tablename, model);
-        return "opened_table";
+        Optional<LinkEntity> linkEntityOptional = linkRepository.findById(currentLink);
+        if (linkEntityOptional.isPresent()) {
+            String tablename = linkEntityOptional.get().getTableEntity().getTablename();
+            linkRepository.deleteById(currentLink);
+            addModelAttributes(tablename, model, new LinkDto());
+            return "opened_table";
+        } else {
+            UserEntity dbUser = userRepository.findById(loggedUsername).get();
+            model.addAttribute("user", dbUser.toDto());
+            model.addAttribute("table", new TableDto());
+            List<TableDto> usertables = dbUser.getTables().stream().map(s -> s.toDto()).collect(Collectors.toList());
+            usertables.sort(new Comparator<TableDto>() {
+                @Override
+                public int compare(TableDto o1, TableDto o2) {
+                    return o1.getTablename().compareToIgnoreCase(o2.getTablename());
+                }
+            });
+            model.addAttribute("usertables", usertables);
+
+            return "logged_user";
+        }
     }
     @Override
     public String addLink(LinkDto linkDto, String tablename, Model model) {
@@ -40,23 +58,32 @@ class LinkServiceImpl implements LinkService {
         linkDto.setId(linkDto.getId().toLowerCase());
 
         if (linkDto.getId().isBlank()) {
-            addModelAttributes(tablename, model);
+            addModelAttributes(tablename, model, linkDto);
             return "blank_linkid";
         } else if (linkDto.getTargetUrl().isBlank()) {
-            addModelAttributes(tablename, model);
+            addModelAttributes(tablename, model, linkDto);
             return "blank_targeturl";
+        } else if (!isValidURL(linkDto)) {
+            addModelAttributes(tablename, model, linkDto);
+            return "wrong_url";
         } else if (linkRepository.existsById(linkDto.getId())) {
-            addModelAttributes(tablename, model);
+            addModelAttributes(tablename, model, linkDto);
             return "duplicated_linkid";
         }
         else {
             LinkEntity newLinkEntity = LinkEntity.fromDto(linkDto);
             newLinkEntity.setTableEntity(tableRepository.findById(tablename).get());
             linkRepository.save(newLinkEntity);
-            addModelAttributes(tablename, model);
+            addModelAttributes(tablename, model, linkDto);
             return "opened_table";
         }
     }
+
+    private boolean isValidURL(LinkDto linkDto) {
+        UrlValidator validator = new UrlValidator();
+        return validator.isValid(linkDto.getTargetUrl());
+    }
+
     @Override
     public String redirectToLinkTables(String tablename, Model model) {
 
@@ -87,7 +114,7 @@ class LinkServiceImpl implements LinkService {
     public String searchForLink(String tablename, KeyWordDto keyWordDto, Model model) {
         String stringToMatch = keyWordDto.getKeyWord().toLowerCase();
         if (stringToMatch.isBlank()) {
-            addModelAttributes(tablename, model);
+            addModelAttributes(tablename, model, new LinkDto());
             return "blank_search";
         }
         List<LinkDto> linklist = tableRepository.findById(tablename).get().getLinks().stream().map(s -> s.toDto()).collect(Collectors.toList());
@@ -112,11 +139,11 @@ class LinkServiceImpl implements LinkService {
 
     @Override
     public String redirectToOpenedTable(String tablename, Model model) {
-        addModelAttributes(tablename, model);
+        addModelAttributes(tablename, model, new LinkDto());
         return "opened_table";
     }
 
-    private void addModelAttributes(String tablename, Model model) {
+    private void addModelAttributes(String tablename, Model model, LinkDto linkDto) {
 
         UserEntity dbUser = tableRepository.findById(tablename).get().getUserEntity();
         List<LinkDto> linklist = tableRepository.findById(tablename).get().getLinks().stream().map(s -> s.toDto()).collect(Collectors.toList());
@@ -129,7 +156,7 @@ class LinkServiceImpl implements LinkService {
         model.addAttribute("user", dbUser.toDto());
         model.addAttribute("linklist", linklist);
         model.addAttribute("tablename", tablename);
-        model.addAttribute("link", new LinkDto());
+        model.addAttribute("link", linkDto);
         model.addAttribute("keyWord", new KeyWordDto());
     }
 }
